pyproject.toml
[tool.poetry]
name = "limerick-bot"
version = "0.1.0"
description = ""
authors = ["Kyle Anthony Williams <kyle.anthony.williams2@gmail.com>"]
readme = "README.md"
packages = [{include = "limerick_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.1.0"
python-dotenv = "^0.21.0"
deep-phonemizer = "^0.0.17"
panphon = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
